requirements.txt
flask
requests
beautifulsoup4
plotly
pandas
jinja2
pytrends
google-generativeai
python-dotenv